add comments to expression
