+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/AnonymousObjectCreationExpression.cs
@@ -14,18 +14,40 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Neo.VM;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertAnonymousObjectCreationExpression(SemanticModel model, AnonymousObjectCreationExpressionSyntax expression)
+    partial class MethodConvert
     {
-        AddInstruction(OpCode.NEWARRAY0);
-        foreach (AnonymousObjectMemberDeclaratorSyntax initializer in expression.Initializers)
+
+        /// <summary>
+        /// Converts C# anonymous type creation expressions to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The anonymous object creation expression syntax node</param>
+        /// <remarks>
+        /// This method handles conversion of expressions such as:
+        ///
+        /// new { X = 10, Y = 20 }
+        ///
+        /// Where an anonymous type is created on the fly and initialized with
+        /// member values.
+        ///
+        /// The equivalent named type syntax would be:
+        ///
+        /// var temp = new MyType { X = 10, Y = 20 };
+        ///
+        /// This method converts the anonymous type syntax so that the
+        /// object creation and member assignments are emitted correctly to executable code.
+        /// </remarks>
+        private void ConvertAnonymousObjectCreationExpression(SemanticModel model, AnonymousObjectCreationExpressionSyntax expression)
         {
-            AddInstruction(OpCode.DUP);
-            ConvertExpression(model, initializer.Expression);
-            AddInstruction(OpCode.APPEND);
+            AddInstruction(OpCode.NEWARRAY0);
+            foreach (AnonymousObjectMemberDeclaratorSyntax initializer in expression.Initializers)
+            {
+                AddInstruction(OpCode.DUP);
+                ConvertExpression(model, initializer.Expression);
+                AddInstruction(OpCode.APPEND);
+            }
         }
     }
 }
