+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/CheckedExpression.cs
@@ -14,14 +14,31 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertCheckedExpression(SemanticModel model, CheckedExpressionSyntax expression)
+    partial class MethodConvert
     {
-        _checkedStack.Push(expression.Keyword.IsKind(SyntaxKind.CheckedKeyword));
-        ConvertExpression(model, expression.Expression);
-        _checkedStack.Pop();
+        /// <summary>
+        /// Converts C# checked expression syntax to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The checked expression syntax node</param>
+        /// <remarks>
+        /// This handles checked syntax like:
+        ///
+        /// checked(x + y)
+        ///
+        /// Where integer overflow checking is explicitly enabled for the expression.
+        ///
+        /// The method will push a "checked" context, convert the inner expression,
+        /// then restore the previous context after. This allows each expression to
+        /// control checking without impacting callers.
+        /// </remarks>
+        private void ConvertCheckedExpression(SemanticModel model, CheckedExpressionSyntax expression)
+        {
+            _checkedStack.Push(expression.Keyword.IsKind(SyntaxKind.CheckedKeyword));
+            ConvertExpression(model, expression.Expression);
+            _checkedStack.Pop();
+        }
     }
 }
