+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/ConditionalExpression.cs
@@ -14,20 +14,39 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Neo.VM;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertConditionalExpression(SemanticModel model, ConditionalExpressionSyntax expression)
+    partial class MethodConvert
     {
-        JumpTarget falseTarget = new();
-        JumpTarget endTarget = new();
-        ConvertExpression(model, expression.Condition);
-        Jump(OpCode.JMPIFNOT_L, falseTarget);
-        ConvertExpression(model, expression.WhenTrue);
-        Jump(OpCode.JMP_L, endTarget);
-        falseTarget.Instruction = AddInstruction(OpCode.NOP);
-        ConvertExpression(model, expression.WhenFalse);
-        endTarget.Instruction = AddInstruction(OpCode.NOP);
+
+        /// <summary>
+        /// Converts C# conditional expressions to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The conditional expression syntax node</param>
+        /// <remarks>
+        /// This handles conversion of ternary conditional expressions like:
+        ///
+        /// result = x > 5 ? 10 : 20;
+        ///
+        /// The condition is first evaluated, followed by a jump to either the true or
+        /// false branch based on the condition value. The respective branch expression is
+        /// evaluated and leaves the final result on the stack.
+        ///
+        /// This allows if-then-else conditional logic in expressions using specialized
+        /// branching instructions.
+        /// </remarks>
+        private void ConvertConditionalExpression(SemanticModel model, ConditionalExpressionSyntax expression)
+        {
+            JumpTarget falseTarget = new();
+            JumpTarget endTarget = new();
+            ConvertExpression(model, expression.Condition);
+            Jump(OpCode.JMPIFNOT_L, falseTarget);
+            ConvertExpression(model, expression.WhenTrue);
+            Jump(OpCode.JMP_L, endTarget);
+            falseTarget.Instruction = AddInstruction(OpCode.NOP);
+            ConvertExpression(model, expression.WhenFalse);
+            endTarget.Instruction = AddInstruction(OpCode.NOP);
+        }
     }
 }
