+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/ElementExpression.cs
@@ -16,76 +16,91 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Neo.VM;
 using System.Linq;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertElementAccessExpression(SemanticModel model, ElementAccessExpressionSyntax expression)
+    partial class MethodConvert
     {
-        if (expression.ArgumentList.Arguments.Count != 1)
-            throw new CompilationException(expression.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {expression.ArgumentList.Arguments}");
-        if (model.GetSymbolInfo(expression).Symbol is IPropertySymbol property)
+        /// <summary>
+        /// Converts array element access expressions to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The element access expression syntax node</param>
+        /// <remarks>
+        /// Handles array access or indexer calls like:
+        ///
+        /// data[i]
+        /// vector[x, y]
+        ///
+        /// Emits code to evaluate array instance and index expression, perform
+        /// bounds checking, and access the element value.
+        /// </remarks>
+        private void ConvertElementAccessExpression(SemanticModel model, ElementAccessExpressionSyntax expression)
         {
-            Call(model, property.GetMethod!, expression.Expression, expression.ArgumentList.Arguments.ToArray());
+            if (expression.ArgumentList.Arguments.Count != 1)
+                throw new CompilationException(expression.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {expression.ArgumentList.Arguments}");
+            if (model.GetSymbolInfo(expression).Symbol is IPropertySymbol property)
+            {
+                Call(model, property.GetMethod!, expression.Expression, expression.ArgumentList.Arguments.ToArray());
+            }
+            else
+            {
+                ITypeSymbol type = model.GetTypeInfo(expression).Type!;
+                ConvertExpression(model, expression.Expression);
+                ConvertIndexOrRange(model, type, expression.ArgumentList.Arguments[0].Expression);
+            }
         }
-        else
+
+        private void ConvertElementBindingExpression(SemanticModel model, ElementBindingExpressionSyntax expression)
         {
+            if (expression.ArgumentList.Arguments.Count != 1)
+                throw new CompilationException(expression.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {expression.ArgumentList.Arguments}");
             ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-            ConvertExpression(model, expression.Expression);
             ConvertIndexOrRange(model, type, expression.ArgumentList.Arguments[0].Expression);
         }
-    }
 
-    private void ConvertElementBindingExpression(SemanticModel model, ElementBindingExpressionSyntax expression)
-    {
-        if (expression.ArgumentList.Arguments.Count != 1)
-            throw new CompilationException(expression.ArgumentList, DiagnosticId.MultidimensionalArray, $"Unsupported array rank: {expression.ArgumentList.Arguments}");
-        ITypeSymbol type = model.GetTypeInfo(expression).Type!;
-        ConvertIndexOrRange(model, type, expression.ArgumentList.Arguments[0].Expression);
-    }
-
-    private void ConvertIndexOrRange(SemanticModel model, ITypeSymbol type, ExpressionSyntax indexOrRange)
-    {
-        if (indexOrRange is RangeExpressionSyntax range)
+        private void ConvertIndexOrRange(SemanticModel model, ITypeSymbol type, ExpressionSyntax indexOrRange)
         {
-            if (range.RightOperand is null)
+            if (indexOrRange is RangeExpressionSyntax range)
             {
-                AddInstruction(OpCode.DUP);
-                AddInstruction(OpCode.SIZE);
+                if (range.RightOperand is null)
+                {
+                    AddInstruction(OpCode.DUP);
+                    AddInstruction(OpCode.SIZE);
+                }
+                else
+                {
+                    ConvertExpression(model, range.RightOperand);
+                }
+                AddInstruction(OpCode.SWAP);
+                if (range.LeftOperand is null)
+                {
+                    Push(0);
+                }
+                else
+                {
+                    ConvertExpression(model, range.LeftOperand);
+                }
+                AddInstruction(OpCode.ROT);
+                AddInstruction(OpCode.OVER);
+                AddInstruction(OpCode.SUB);
+                switch (type.ToString())
+                {
+                    case "byte[]":
+                        AddInstruction(OpCode.SUBSTR);
+                        break;
+                    case "string":
+                        AddInstruction(OpCode.SUBSTR);
+                        ChangeType(VM.Types.StackItemType.ByteString);
+                        break;
+                    default:
+                        throw new CompilationException(indexOrRange, DiagnosticId.ArrayRange, $"The type {type} does not support range access.");
+                }
             }
             else
             {
-                ConvertExpression(model, range.RightOperand);
+                ConvertExpression(model, indexOrRange);
+                AddInstruction(OpCode.PICKITEM);
             }
-            AddInstruction(OpCode.SWAP);
-            if (range.LeftOperand is null)
-            {
-                Push(0);
-            }
-            else
-            {
-                ConvertExpression(model, range.LeftOperand);
-            }
-            AddInstruction(OpCode.ROT);
-            AddInstruction(OpCode.OVER);
-            AddInstruction(OpCode.SUB);
-            switch (type.ToString())
-            {
-                case "byte[]":
-                    AddInstruction(OpCode.SUBSTR);
-                    break;
-                case "string":
-                    AddInstruction(OpCode.SUBSTR);
-                    ChangeType(VM.Types.StackItemType.ByteString);
-                    break;
-                default:
-                    throw new CompilationException(indexOrRange, DiagnosticId.ArrayRange, $"The type {type} does not support range access.");
-            }
-        }
-        else
-        {
-            ConvertExpression(model, indexOrRange);
-            AddInstruction(OpCode.PICKITEM);
         }
     }
 }
