+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/IdentifierNameExpression.cs
@@ -16,62 +16,84 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Neo.VM;
 using System;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertIdentifierNameExpression(SemanticModel model, IdentifierNameSyntax expression)
+    partial class MethodConvert
     {
-        ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
-        switch (symbol)
+        /// <summary>
+        /// Converts identifier name expressions to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The identifier expression syntax node</param>
+        /// <remarks>
+        /// This handles identifiers referring to fields, locals, parameters,
+        /// properties etc.
+        ///
+        /// Examples:
+        ///
+        /// myLocal
+        /// MyField
+        /// obj.Prop
+        ///
+        /// The method determines the target symbol and emits instructions
+        /// to load/access the value based on its type.
+        ///
+        /// This includes handling constant values, static vs instance
+        /// members, properties and more.
+        /// </remarks>
+        private void ConvertIdentifierNameExpression(SemanticModel model, IdentifierNameSyntax expression)
         {
-            case IFieldSymbol field:
-                if (field.IsConst)
-                {
-                    Push(field.ConstantValue);
-                }
-                else if (field.IsStatic)
-                {
-                    byte index = context.AddStaticField(field);
-                    AccessSlot(OpCode.LDSFLD, index);
-                }
-                else
-                {
-                    int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                    AddInstruction(OpCode.LDARG0);
-                    Push(index);
-                    AddInstruction(OpCode.PICKITEM);
-                }
-                break;
-            case ILocalSymbol local:
-                if (local.IsConst)
-                    Push(local.ConstantValue);
-                else
-                    AccessSlot(OpCode.LDLOC, _localVariables[local]);
-                break;
-            case IMethodSymbol method:
-                if (!method.IsStatic)
-                    throw new CompilationException(expression, DiagnosticId.NonStaticDelegate, $"Unsupported delegate: {method}");
-                MethodConvert convert = context.ConvertMethod(model, method);
-                Jump(OpCode.PUSHA, convert._startTarget);
-                break;
-            case IParameterSymbol parameter:
-                if (!_internalInline)
-                    AccessSlot(OpCode.LDARG, _parameters[parameter]);
-                break;
-            case IPropertySymbol property:
-                if (property.IsStatic)
-                {
-                    Call(model, property.GetMethod!);
-                }
-                else
-                {
-                    AddInstruction(OpCode.LDARG0);
-                    Call(model, property.GetMethod!);
-                }
-                break;
-            default:
-                throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
+            ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
+            switch (symbol)
+            {
+                case IFieldSymbol field:
+                    if (field.IsConst)
+                    {
+                        Push(field.ConstantValue);
+                    }
+                    else if (field.IsStatic)
+                    {
+                        byte index = context.AddStaticField(field);
+                        AccessSlot(OpCode.LDSFLD, index);
+                    }
+                    else
+                    {
+                        int index = Array.IndexOf(field.ContainingType.GetFields(), field);
+                        AddInstruction(OpCode.LDARG0);
+                        Push(index);
+                        AddInstruction(OpCode.PICKITEM);
+                    }
+                    break;
+                case ILocalSymbol local:
+                    if (local.IsConst)
+                        Push(local.ConstantValue);
+                    else
+                        AccessSlot(OpCode.LDLOC, _localVariables[local]);
+                    break;
+                case IMethodSymbol method:
+                    if (!method.IsStatic)
+                        throw new CompilationException(expression, DiagnosticId.NonStaticDelegate, $"Unsupported delegate: {method}");
+                    MethodConvert convert = context.ConvertMethod(model, method);
+                    Jump(OpCode.PUSHA, convert._startTarget);
+                    break;
+                case IParameterSymbol parameter:
+                    if (!_internalInline)
+                        AccessSlot(OpCode.LDARG, _parameters[parameter]);
+                    break;
+                case IPropertySymbol property:
+                    if (property.IsStatic)
+                    {
+                        Call(model, property.GetMethod!);
+                    }
+                    else
+                    {
+                        AddInstruction(OpCode.LDARG0);
+                        Call(model, property.GetMethod!);
+                    }
+                    break;
+                default:
+                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
+            }
         }
     }
 }
