+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/InterpolatedStringExpression.cs
@@ -14,41 +14,57 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Neo.VM;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertInterpolatedStringExpression(SemanticModel model, InterpolatedStringExpressionSyntax expression)
+    partial class MethodConvert
     {
-        if (expression.Contents.Count == 0)
+        /// <summary>
+        /// Converts interpolated string expressions to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The interpolated string syntax</param>
+        /// <remarks>
+        /// Handles string syntax like:
+        ///
+        /// $"Hello {name}"
+        /// $"Value is {value}"
+        ///
+        /// The contents are parsed to extract literal text
+        /// and converted expressions. Emitted code concatenates
+        /// them together correctly into the final string value.
+        /// </remarks>
+        private void ConvertInterpolatedStringExpression(SemanticModel model, InterpolatedStringExpressionSyntax expression)
         {
-            Push(string.Empty);
-            return;
+            if (expression.Contents.Count == 0)
+            {
+                Push(string.Empty);
+                return;
+            }
+            ConvertInterpolatedStringContent(model, expression.Contents[0]);
+            for (int i = 1; i < expression.Contents.Count; i++)
+            {
+                ConvertInterpolatedStringContent(model, expression.Contents[i]);
+                AddInstruction(OpCode.CAT);
+            }
+            if (expression.Contents.Count >= 2)
+                ChangeType(VM.Types.StackItemType.ByteString);
         }
-        ConvertInterpolatedStringContent(model, expression.Contents[0]);
-        for (int i = 1; i < expression.Contents.Count; i++)
-        {
-            ConvertInterpolatedStringContent(model, expression.Contents[i]);
-            AddInstruction(OpCode.CAT);
-        }
-        if (expression.Contents.Count >= 2)
-            ChangeType(VM.Types.StackItemType.ByteString);
-    }
 
-    private void ConvertInterpolatedStringContent(SemanticModel model, InterpolatedStringContentSyntax content)
-    {
-        switch (content)
+        private void ConvertInterpolatedStringContent(SemanticModel model, InterpolatedStringContentSyntax content)
         {
-            case InterpolatedStringTextSyntax syntax:
-                Push(syntax.TextToken.ValueText);
-                break;
-            case InterpolationSyntax syntax:
-                if (syntax.AlignmentClause is not null)
-                    throw new CompilationException(syntax.AlignmentClause, DiagnosticId.AlignmentClause, $"Alignment clause is not supported: {syntax.AlignmentClause}");
-                if (syntax.FormatClause is not null)
-                    throw new CompilationException(syntax.FormatClause, DiagnosticId.FormatClause, $"Format clause is not supported: {syntax.FormatClause}");
-                ConvertObjectToString(model, syntax.Expression);
-                break;
+            switch (content)
+            {
+                case InterpolatedStringTextSyntax syntax:
+                    Push(syntax.TextToken.ValueText);
+                    break;
+                case InterpolationSyntax syntax:
+                    if (syntax.AlignmentClause is not null)
+                        throw new CompilationException(syntax.AlignmentClause, DiagnosticId.AlignmentClause, $"Alignment clause is not supported: {syntax.AlignmentClause}");
+                    if (syntax.FormatClause is not null)
+                        throw new CompilationException(syntax.FormatClause, DiagnosticId.FormatClause, $"Format clause is not supported: {syntax.FormatClause}");
+                    ConvertObjectToString(model, syntax.Expression);
+                    break;
+            }
         }
     }
 }
