+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/IsPatternExpression.cs
@@ -14,16 +14,29 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Neo.VM;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertIsPatternExpression(SemanticModel model, IsPatternExpressionSyntax expression)
+    partial class MethodConvert
     {
-        byte anonymousIndex = AddAnonymousVariable();
-        ConvertExpression(model, expression.Expression);
-        AccessSlot(OpCode.STLOC, anonymousIndex);
-        ConvertPattern(model, expression.Pattern, anonymousIndex);
-        RemoveAnonymousVariable(anonymousIndex);
+
+        /// <summary>
+        /// Converts 'is' pattern expression syntax to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The 'is' pattern expression syntax</param>
+        /// <remarks>
+        /// Handles syntax like:
+        ///
+        /// expr is Type name
+        ///
+        /// </remarks>
+        private void ConvertIsPatternExpression(SemanticModel model, IsPatternExpressionSyntax expression)
+        {
+            byte anonymousIndex = AddAnonymousVariable();
+            ConvertExpression(model, expression.Expression);
+            AccessSlot(OpCode.STLOC, anonymousIndex);
+            ConvertPattern(model, expression.Pattern, anonymousIndex);
+            RemoveAnonymousVariable(anonymousIndex);
+        }
     }
 }
