+++ src/Neo.Compiler.CSharp/MethodConvert/Expression/MemberExpression.cs
@@ -16,63 +16,78 @@ using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Neo.VM;
 using System;
 
-namespace Neo.Compiler;
-
-partial class MethodConvert
+namespace Neo.Compiler
 {
-    private void ConvertMemberAccessExpression(SemanticModel model, MemberAccessExpressionSyntax expression)
+    partial class MethodConvert
     {
-        ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
-        switch (symbol)
+        /// <summary>
+        /// Converts member access expressions to executable code.
+        /// </summary>
+        /// <param name="model">The semantic model</param>
+        /// <param name="expression">The member access expression syntax</param>
+        /// <remarks>
+        /// Handles expressions accessing fields, properties etc like:
+        ///
+        /// obj.Field
+        /// Class.StaticProp
+        ///
+        /// Determines symbol type and emits instructions to load/access
+        /// the value appropriately. Handles static vs instance members.
+        /// </remarks>
+        private void ConvertMemberAccessExpression(SemanticModel model, MemberAccessExpressionSyntax expression)
         {
-            case IFieldSymbol field:
-                if (field.IsConst)
-                {
-                    Push(field.ConstantValue);
-                }
-                else if (field.IsStatic)
-                {
-                    byte index = context.AddStaticField(field);
-                    AccessSlot(OpCode.LDSFLD, index);
-                }
-                else
-                {
+            ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
+            switch (symbol)
+            {
+                case IFieldSymbol field:
+                    if (field.IsConst)
+                    {
+                        Push(field.ConstantValue);
+                    }
+                    else if (field.IsStatic)
+                    {
+                        byte index = context.AddStaticField(field);
+                        AccessSlot(OpCode.LDSFLD, index);
+                    }
+                    else
+                    {
+                        int index = Array.IndexOf(field.ContainingType.GetFields(), field);
+                        ConvertExpression(model, expression.Expression);
+                        Push(index);
+                        AddInstruction(OpCode.PICKITEM);
+                    }
+                    break;
+                case IMethodSymbol method:
+                    if (!method.IsStatic)
+                        throw new CompilationException(expression, DiagnosticId.NonStaticDelegate, $"Unsupported delegate: {method}");
+                    MethodConvert convert = context.ConvertMethod(model, method);
+                    Jump(OpCode.PUSHA, convert._startTarget);
+                    break;
+                case IPropertySymbol property:
+                    ExpressionSyntax? instanceExpression = property.IsStatic ? null : expression.Expression;
+                    Call(model, property.GetMethod!, instanceExpression);
+                    break;
+                default:
+                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
+            }
+        }
+
+        private void ConvertMemberBindingExpression(SemanticModel model, MemberBindingExpressionSyntax expression)
+        {
+            ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
+            switch (symbol)
+            {
+                case IFieldSymbol field:
                     int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                    ConvertExpression(model, expression.Expression);
                     Push(index);
                     AddInstruction(OpCode.PICKITEM);
-                }
-                break;
-            case IMethodSymbol method:
-                if (!method.IsStatic)
-                    throw new CompilationException(expression, DiagnosticId.NonStaticDelegate, $"Unsupported delegate: {method}");
-                MethodConvert convert = context.ConvertMethod(model, method);
-                Jump(OpCode.PUSHA, convert._startTarget);
-                break;
-            case IPropertySymbol property:
-                ExpressionSyntax? instanceExpression = property.IsStatic ? null : expression.Expression;
-                Call(model, property.GetMethod!, instanceExpression);
-                break;
-            default:
-                throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
-        }
-    }
-
-    private void ConvertMemberBindingExpression(SemanticModel model, MemberBindingExpressionSyntax expression)
-    {
-        ISymbol symbol = model.GetSymbolInfo(expression).Symbol!;
-        switch (symbol)
-        {
-            case IFieldSymbol field:
-                int index = Array.IndexOf(field.ContainingType.GetFields(), field);
-                Push(index);
-                AddInstruction(OpCode.PICKITEM);
-                break;
-            case IPropertySymbol property:
-                Call(model, property.GetMethod!);
-                break;
-            default:
-                throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
+                    break;
+                case IPropertySymbol property:
+                    Call(model, property.GetMethod!);
+                    break;
+                default:
+                    throw new CompilationException(expression, DiagnosticId.SyntaxNotSupported, $"Unsupported symbol: {symbol}");
+            }
         }
     }
 }
